requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
pymupdf
python-docx
chromadb>=0.4.22
sentence-transformers
pinecone 